Don't try to use the jemalloc implementation with Rust on seL4.
Using jemalloc seems to trigger inclusion of a number of stack checking code, which in turn try to use thread-local storage. This causes a crash, because seL4 doesn't set up `fs` and `gs` for TLS in exactly the same way as Linux. If you see a crash in `__init_ssp`, then you're running into this problem.